+++ R/Sojourn.R
@@ -33,3 +33,4 @@ NULL
 #' @import nnet
 NULL
 
+if(getRversion() >= "2.15.1")  utils::globalVariables(c("."))