+++ R/soj_1x_original.R
@@ -1,6 +1,6 @@
 #' Invoke the original uni-axial Sojourn method
 #'
-#' Calls the uni-axial Sojourn method from \href{https://www.ncbi.nlm.nih.gov/pubmed/23860415}{Lyden et al. (2014)}.
+#' Calls the uni-axial Sojourn method from \href{https://pubmed.ncbi.nlm.nih.gov/23860415/}{Lyden et al. (2014)}.
 #'
 #' @param counts numeric vector of vertical axis counts
 #' @param perc.cut decision tree threshold 1
@@ -15,18 +15,19 @@
 #'
 #' @examples
 #' data(example_data, package = "Sojourn")
-#' results_1x <- soj_1x_original(example_data$axis1)
-#'
-#' head(results_1x)
-#'
+#' if (isTRUE(requireNamespace("Sojourn.Data"))) {
+#'   results_1x <- soj_1x_original(example_data$axis1)
+#'   utils::head(results_1x)
+#' }
 soj_1x_original <- function(counts,perc.cut=0.05,perc.cut.2=0.12,
   perc.cut.3=0.55,too.short=10,sit.cut=90,long.soj=120) {
 
   if (!requireNamespace("Sojourn.Data", quietly = TRUE)) {
-    stop(paste(
+    stop(
       "You must install the package `Sojourn.Data`",
-      "to use this function."
-    ))
+      "to use this function.\n  If it is missing on CRAN, use ",
+      "devtools::install_github(\"paulhibbing/Sojourn.Data\")"
+    )
   }
 
   y <- counts